+++ src/Library/Comunicadores/GuardadoNotificacion.cs
@@ -7,7 +7,7 @@ namespace Library
     /// </summary>
     public class GuardadoNotificacion: ConfiguracionNotificacion
     {
-        public GuardadoNotificacion(string mensajeEntrada, int iDUsuario):base(mensajeEntrada,iDUsuario){}
+        public GuardadoNotificacion(string mensajeEntrada, long iDUsuario):base(mensajeEntrada,iDUsuario){}
 
         public override void Manipular()
         {
